// src/app/views/common/screen-resolution/util/resolution-types.ts
export enum ScreenResolution {
  MOBILE = 'mobile',
  TABLET = 'tablet',
  TABLET_MEDIUM = 'tablet_medium',
  TABLET_LARGE = 'tablet_large',
  LAPTOP_SMALL = 'laptop_small',
  LAPTOP_MEDIUM = 'laptop_medium',
  LAPTOP_LARGE = 'laptop_large',
  DESKTOP = 'desktop',
  DESKTOP_LARGE = 'desktop_large',
  DESKTOP_XLARGE = 'desktop_xlarge',
  DESKTOP_XXLARGE = 'desktop_xxlarge'
}